Overhaul communication.
Cah_bot assumes that all commands come over channel that the game is being
played in. If a player instead gives commands over PM, this can sometimes cause
messages which are intended for the whole channel - like the list of answers
for judging - to be sent to just one player.

This commit addresses that problem by splitting all communication into three
classes:

    - Announcements: Messages that are relevant to all players. These are
      always sent to the channel.

    - Private messages: Messages containing secret information like a player's
      hand. These are always sent directly to the player via a NOTICE.

    - Responses to a command, error messages, and anything that doesn't fall
      into the above categories. Sent the same way as the inciting command.

To be able to send announcements at any time, cah_bot needs to know which
channel the game is in. As a temporary measure, this is accomplished via a
configuration variable. A later commit will add some smarts to the !join
command instead.